requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
pymupdf4llm
arxiv
duckduckgo-search
newspaper3k
plotly
supabase
lxml_html_clean
python-dotenv
requests
sentence-transformers
langchain-text-splitters
streamlit
langchain-groq
fpdf
python-docx
